projects/project1-learning-chatbot/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM APIs
openai==1.6.0
anthropic==0.7.7
google-generativeai==0.3.0
langchain==0.0.350
langchain-community==0.0.10

# File processing
PyPDF2==3.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==1.1.0
markdown==3.5.1

# Text processing
nltk==3.8.1
spacy==3.7.2
transformers==4.36.0
sentence-transformers==2.2.2

# Vector database
chromadb==0.4.18
pinecone-client==2.2.4

# Utilities
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0